+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-> ButtonGroup className 修改为 `bfd-btn-group`
-
 - DataTable onRowClick 新增 index、event 参数
 - TreeSelect 新增属性 dataFilter
-- DataTable2 新增属性 rowRender，自定义行渲染
+- DataTable2 新增属性 rowRender，自定义行渲染；无数据显示处理；
 - 修复 Select 无匹配时样式问题
 - ButtonGroup 新增属性 value，修改为 inline-block 模式
+- Fetch 新增 reload 接口，可强制重新加载
+- 修复 TreeSelect value 为0时的问题
+- 修正了部分组件的样式
 
 
 ## v1.3.0